Refactor WRFState and remove VariableGridInfo structure
- Replaced the inclusion of Location.hpp with WRFGridInfo.hpp in WRFState.hpp.
- Removed the VariableGridInfo struct and its associated methods, streamlining the WRFState class for improved clarity and maintainability.
- Updated documentation to reflect the changes in grid information handling, enhancing overall code organization.
